policyeval: Implement MSC4284 policy server (#21)
This PR implements a standalone [policy
server](https://github.com/matrix-org/matrix-spec-proposals/pull/4284)
which fully supports and has been tested using the Synapse
[policyserv_spam_checker](https://github.com/element-hq/policyserv_spam_checker)
module.

To set it up, all you need to do is pass the
`/_matrix/policy/unstable/org.matrix.msc4284/event/{event_id}/check`
endpoint through to Meowlnir in the reverse proxy. Meowlnir will work
out which policies to apply based on the room the PDU was sent in.
Optionally, Meowlnir can be configured to send automatic redactions for
events that fail the check.
Meowlnir must be joined to, and the PDU sent in a protected room, in
order for this to work.

An example server has been deployed at nexy7574.co.uk, and has been
working seamlessly in my rooms so far.

---------

Co-authored-by: Tulir Asokan <tulir@maunium.net>