policyserver: Selectively dispatch redactions (#31)
This change tells the policy server to only cache the result and
conditionally issue the relevant redaction **only** when the server
calling /check **is not the one sending the event**. If `example.com`
calls `/check` on an event sent by `@foo:example.com`, it is assumed to
be rejected locally, preventing it from ever reaching the timeline,
meaning a followup redaction would be pointless.

If the server calls the check, but still federates the event, another
server is likely to call /check on it anyway, resulting in the regular
redaction flow being triggered.